% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadAxfBOM.R
\name{ReadAxfBOM}
\alias{ReadAxfBOM}
\title{Import a weather station observation axf-file downloaded from a Australian Bureau of Meteorology station}
\usage{
ReadAxfBOM(data)
}
\arguments{
\item{data}{Filename of the axf file.}
}
\value{
Data frame with the observations in file. See http://www.bom.gov.au/weather-services/about/IDY03000.doc for details
}
\description{
Import a weather station observation axf-file downloaded from a Australian Bureau of Meteorology station
}
\examples{
\dontrun{
observations <- ReadAxfBOM("filename.axf")
}
}

